# backend/app/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/admin/memberships":{"get":{"summary":"List Memberships","operationId":"list_memberships_admin_memberships_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/MembershipOut"},"type":"array","title":"Response List Memberships Admin Memberships Get"}}}}}},"post":{"summary":"Create Membership","operationId":"create_membership_admin_memberships_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MembershipCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MembershipOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/assign":{"post":{"summary":"Assign Membership","operationId":"assign_membership_admin_assign_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserMembershipAssignRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserMembershipOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/user/debug":{"get":{"summary":"Debug Memberships","operationId":"debug_memberships_admin_user_debug_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/user/membership":{"get":{"summary":"Get User Membership","operationId":"get_user_membership_user_membership_get","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserMembershipStatusOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/use-feature":{"post":{"summary":"Use Feature","operationId":"use_feature_user_use_feature_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UseFeatureRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UseFeatureResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"FeatureStatus":{"properties":{"feature_name":{"type":"string","title":"Feature Name"},"used_count":{"type":"integer","title":"Used Count"},"remaining_count":{"type":"integer","title":"Remaining Count"}},"type":"object","required":["feature_name","used_count","remaining_count"],"title":"FeatureStatus"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MembershipCreate":{"properties":{"name":{"type":"string","title":"Name"},"duration_days":{"type":"integer","title":"Duration Days"},"features":{"items":{"$ref":"#/components/schemas/MembershipFeatureCreate"},"type":"array","title":"Features"}},"type":"object","required":["name","duration_days","features"],"title":"MembershipCreate"},"MembershipFeatureCreate":{"properties":{"feature_name":{"type":"string","enum":["chat","analysis"],"title":"Feature Name"},"limit_count":{"type":"integer","title":"Limit Count"}},"type":"object","required":["feature_name","limit_count"],"title":"MembershipFeatureCreate"},"MembershipFeatureOut":{"properties":{"feature_name":{"type":"string","enum":["chat","analysis"],"title":"Feature Name"},"limit_count":{"type":"integer","title":"Limit Count"},"id":{"type":"string","title":"Id"}},"type":"object","required":["feature_name","limit_count","id"],"title":"MembershipFeatureOut"},"MembershipOut":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"duration_days":{"type":"integer","title":"Duration Days"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"features":{"items":{"$ref":"#/components/schemas/MembershipFeatureOut"},"type":"array","title":"Features"}},"type":"object","required":["id","name","duration_days","created_at","features"],"title":"MembershipOut"},"UseFeatureRequest":{"properties":{"user_id":{"type":"string","title":"User Id"},"feature_name":{"type":"string","title":"Feature Name"}},"type":"object","required":["user_id","feature_name"],"title":"UseFeatureRequest"},"UseFeatureResponse":{"properties":{"success":{"type":"boolean","title":"Success"},"message":{"type":"string","title":"Message"},"remaining_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Remaining Count"}},"type":"object","required":["success","message"],"title":"UseFeatureResponse"},"UserMembershipAssignRequest":{"properties":{"user_id":{"type":"string","title":"User Id"},"membership_id":{"type":"string","format":"uuid","title":"Membership Id"}},"type":"object","required":["user_id","membership_id"],"title":"UserMembershipAssignRequest"},"UserMembershipOut":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"user_id":{"type":"string","title":"User Id"},"membership_id":{"type":"string","format":"uuid","title":"Membership Id"},"start_date":{"type":"string","format":"date-time","title":"Start Date"},"end_date":{"type":"string","format":"date-time","title":"End Date"}},"type":"object","required":["id","user_id","membership_id","start_date","end_date"],"title":"UserMembershipOut"},"UserMembershipStatusOut":{"properties":{"membership_name":{"type":"string","title":"Membership Name"},"start_date":{"type":"string","format":"date-time","title":"Start Date"},"end_date":{"type":"string","format":"date-time","title":"End Date"},"features":{"items":{"$ref":"#/components/schemas/FeatureStatus"},"type":"array","title":"Features"}},"type":"object","required":["membership_name","start_date","end_date","features"],"title":"UserMembershipStatusOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}